+++ lib/modules/admin-widgets/lib/api.js
@@ -24,7 +24,7 @@ module.exports = (self, options) => {
     const mappedIdeas = ideas.map((idea) => {
       return {
         'titel': idea.title,
-        'locatie': idea.location.coordinates[0] + ', ' + idea.location.coordinates[1],
+        'locatie':idea.location && idea.location.coordinates ? idea.location.coordinates[0] + ', ' + idea.location.coordinates[1] : '',
         'aangemaakt': moment(idea.createdAt).format('DD-MM-YYYY HH:mm'),
         'thema': idea.extraData.theme || ''
       }
