+++ lib/modules/idea-overview-widgets/views/modbreak.html
@@ -4,7 +4,7 @@
 			<b>{{data.global.modbreakAuthor}} | </b>
 		{% endif %}
 
-		<b>{{idea.modBreakDateNice}}</b>
+		<b>{{idea.modBreakDate | date( 'D MMM YYYY, HH:mm')}}</b>
 		<br>
 		<p>
 			{{idea.modBreak | stripSafe | safe | nlbr}}
