+++ lib/modules/idea-single-widgets/views/includes/share-idea-buttons.html
@@ -3,6 +3,6 @@
 	<ul>
 		<li><a class="facebook" target="_blank" href="https://www.facebook.com/sharer/sharer.php?u={{url}}">Facebook</a></li>
 		<li><a class="twitter"  target="_blank" href="https://twitter.com/intent/tweet?text={{url}}">Twitter</a></li>
-		<li><a class="email"    target="_blank" href="mailto:?subject={{'Idee voor West Begroot: '+idea.title | urlencode}}&body={{url | urlencode}}">Email</a></li>
+		<li><a class="email"    target="_blank" href="mailto:?subject={{idea.title | urlencode}}&body={{url | urlencode}}">Email</a></li>
 	</ul>
 </div>
