+++ lib/modules/participatory-budgeting-widgets/public/css/budget-block.less
