+++ lib/modules/participatory-budgeting-widgets/public/css/button-vote.less
