+++ lib/modules/participatory-budgeting-widgets/public/css/end-date-bar.less
