+++ lib/modules/participatory-budgeting-widgets/public/css/links.less
