+++ lib/modules/participatory-budgeting-widgets/public/css/mobile-accordion.less
