+++ lib/modules/participatory-budgeting-widgets/public/css/overview.less
