+++ lib/modules/participatory-budgeting-widgets/public/css/sticky.less
