+++ lib/modules/participatory-budgeting-widgets/public/js/ideas-lister.js
