+++ lib/modules/participatory-budgeting-widgets/public/js/voting.js
