+++ lib/modules/participatory-budgeting-widgets/views/phase-voting/ideas-list.html
