+++ lib/modules/participatory-budgeting-widgets/views/phase-voting/includes/end-date-bar.njk
