+++ locales/default.json
@@ -384,5 +384,40 @@
 	"Display counter": "Display counter",
 	"Counter text": "Counter text",
 	"Edit Ideas map": "Edit Ideas map",
-	"You are now editing this document in another tab or window.": "You are now editing this document in another tab or window."
+	"You are now editing this document in another tab or window.": "You are now editing this document in another tab or window.",
+	"Uitloggen": "Uitloggen",
+	"Edit Idea single": "Edit Idea single",
+	"Extra info": "Extra info",
+	"Phone number": "Phone number",
+	"Firstname": "Firstname",
+	"Lastname": "Lastname",
+	"Edit Arguments form": "Edit Arguments form",
+	"Google Tag Manager Property ID (like GTM-xxxxx)": "Google Tag Manager Property ID (like GTM-xxxxx)",
+	"Url": "Url",
+	"Open in new tab": "Open in new tab",
+	"Save Page header": "Save Page header",
+	"Subtitle": "Subtitle",
+	"Intro text": "Intro text",
+	"Show button(s) under intro text": "Show button(s) under intro text",
+	"Left button label": "Left button label",
+	"Left button URL": "Left button URL",
+	"Right button label": "Right button label",
+	"Right button URL": "Right button URL",
+	"Show button in right column": "Show button in right column",
+	"Button label": "Button label",
+	"Button URL": "Button URL",
+	"Idea count is fixed": "Idea count is fixed",
+	"Fixed idea count (max 3 digits)": "Fixed idea count (max 3 digits)",
+	"Edit Page header": "Edit Page header",
+	"Save List": "Save List",
+	"Select type of list": "Select type of list",
+	"Unordered list": "Unordered list",
+	"Ordered list": "Ordered list",
+	"Select appearance modus for list": "Select appearance modus for list",
+	"Bullets": "Bullets",
+	"Checkmarks with blue background": "Checkmarks with blue background",
+	"Blue Checkmarks": "Blue Checkmarks",
+	"Stripes": "Stripes",
+	"Numbers": "Numbers",
+	"Mijn account": "Mijn account"
 }
